Add support for env and fatalError options
env allows to specify environment variables that will be passed to the
command being executed.

fatalError allows to change the default behaviour that will abort the
test execution, allowing for expected failures to not break k6.

When the command exits with a non zero error code, the stderr will be
returned instead of the stdout